Temporary patch: ensure errors lead to exit(1) in main funcs (#6318)
The main docker-compose.yml is currently erroring silently without printing anything due to... either the urfave/cli/v2 migration (i.e. changes in urfave), or simply a long-standing silent-failure bug that should be squashed.
So this changes all our `func main`s to check for error returns, and print and exit them.

We're going to be doing some rewriting of error handling in these CLIs very soon, so for now I'm just doing the simplest patch possible, to ease troubleshooting.