Feature/zonal isolation zone discovery (#6301)
This changes the config property system.AllIsolationGroups from being a fixed static configuration value loaded on startup to be a callback which can be updated more dynamically. This should continue working with the existing dynamic config property as before, but support passing in a more complex discovery mechanism for isolation groups as needed.

Why?

We found that we were having trouble updating this config in production and needed to refactor it slightly to allow for a better update system.

How did you test it?

Deployed in staging environments with some initial manual testing as well as the unit testing.

Potential risks

This could break zonal isolation-the feature if it is wrong or buggy. that feature is designed to fall back to tasks being simply un-isolated, so it's not expected to actually break task processing, but it could degrade it if it were to contain some unforseen problems.