Squash NaN bugs and prevent them from coming back.
Addresses some rare (but damaging) issues in production that appear to be due to NaNs in
aggregator responses, very likely due to sending 0-rps updates at unlucky times and/or
update-ticks running more quickly than 1s (which may occur if an update takes >2s, as
tickers queue up a "tick" and their time continues elapsing whether it is read or not).

These NaNs do two very problematic things:
- NaNs beget more NaNs, so they tend to "stick" for large periods of time
- NaNs turn into `INT_MIN` which related code tends to treat as `0` (e.g. the underlying ratelimiter does this for negative limits)
  - and this rejects all requests

---

This level of check-paranoia is almost certainly excessive, but:
- this would have prevented some rare but significant problems
- bad-data fallbacks exist (it's just a failed update) and these paths should use it instead of misbehaving
- this kind of issue is tough to spot during code review, so tests (fuzz in particular) are likely necessary
- unexpected values in this kind of math can be very hard to track down because they cause problems many lines later,
  so logging as early as possible can help a lot.

... so I've perhaps gone a bit overboard, but the end result seems fairly reasonable I think?
And hopefully resistant to future mistakes, though that does somewhat rely on people running the fuzz tests when they make changes.

---

Fuzz testing specifically, I've included some of the failed runs while I was developing it, to seed the fuzzer,
and these are run as part of normal `go test` runs so they also help ensure coverage.

Some of the scenarios took ~20 minutes of fuzzing to catch though, so as is normal with fuzz testing:
If you make changes and you are not totally confident in them, fuzzing for *hours* is sorta the baseline expectation.