Refactored and tested SerializeForLogging (#6456)
What changed?
Refactored SerializeForLogging so we use a wrapper instead of implementing it on every single request type. The function is used to give people implementing custum authorizers the ability to log the request body.

We supported nil receivers in the old interface, so to not break this support we here support typed nils.

Why?
Improve code quality and test coverage

How did you test it?

Potential risks

Release notes

Documentation Changes