Change Isolation Group Fallback Behavior
Change default-partitioner to fall back to any group if the requested group isn't available rather than attempting to pick a stable zone based on the workflow ID. This optimizes for lower task latency and allows for better recovery from skew or drain scenarios.

Allow for tasks to fall back from the requested zone outside of the root partition. If workflows are started in an isolation group where no pollers are present, only the root partition will allow those tasks to leak to other isolation groups. This bottlenecks the rate at which those tasks can be completed by the rate at which we forward tasks to the root partition.

Change task_reader.go to recalculate the isolation group of each task in the buffer before dispatching it. Currently taskReader calculates the isolation group of each task when it reads it from persistence, places it in the buffer for that isolation group, and then dispatches it using that same isolation group. When an isolation group becomes unavailable any buffered tasks will need to experience a dispatch timeout before they're reassigned. This will happen sequentially, dispatching tasks at a rate of 1 task per 3 seconds (async dispatch timeout). This is particularly problematic when combined with the startup behavior that considers all isolation groups to be present. The only relief from this behavior is forwarding tasks to the root partition, which will incidentally recalculate the isolation group of the task when it dispatches it. In the future, this will also be important for supporting task isolation duration.

Since the returned group is now guaranteed to be either the original group for the task or the default buffer, we can remove the logic for copying tasks between buffers upon group reassignment. Tasks previously could be reassigned to any zone and needed to be copied to the buffer for that zone to ensure that attempting to dispatch in that zone didn't block dispatching any other buffered tasks for the original zone.

Additionally include some refactoring to allow for a task isolation duration in the future.