Rename new shardmanager service to shard-distributor (#6496)
What changed?
Renamed the new shard-manager service to shard-distributor

Why?
We want to split the shard management system in two service, shard-distributor and a load-balancer. Also there is already a shard manager struct in the code, which would be quite confusing

How did you test it?
Unit tests

Potential risks
The change is NOT backwards compatible, the service is still prototyping so noone show rely on it. We need to change the names internally when we pull in this change.

Release notes

Documentation Changes