Bugfix/picking changes to allow for failover polling (#6523)
With thanks to my colleague @Shaddoll for noticing, this ports across a Temporal fix temporalio/temporal#4943 to the cadence server repo.

The problem is that for sync workflows, or workflows using the Execute API in go, the branch token itself is not sufficiently good enough information about the state of history polling. It can change, such as during a failover. This wraps that change in version-history so that it's able to handle this more gracefully.

Why?
To fix a problem for customers during failover