Update weighted load balancer (#6522)
## What?
The main purpose of this PR is to update weighted load balancer to support activity task list. A new proto type is introduced for this purpose and some refactoring is done to make the code more readable.

## Test
This PR has unit tests for all components. The end-to-end tests was done manually by running workflows in dev2.