Add testing.T arg to test-only helpers
mutable_stat_util.go is surprisingly mixed in its purpose, though it is not documented as such in any way I can see.

So, since these kinda-dubious copy-helpers are in here but are not clearly test-only, I've added a testing.T arg to make it clear and prevent their use elsewhere.
More cleanup should happen in this file, but they're all quite semantically different so I'm breaking it up.