Split mutable_state_util.go by purpose (#6577)
It's rather confusing that this file is half test-only helpers and half "used TONS of locations" utils.
So let's split it up.

Since these are used in multiple packages, I can't make them private, and they also use private things in this package so I can't move them out.
So it's just split.  Not ideal but an improvement at least.